Show fat and skinny metadata extents
The Btrfs trees are stored as metadata inside extents. The extents used
for metadata are tracked inside the metadata. Crazy!

Extents could already been shown since a66853e "Show extents inside a
block group". This was extended with showing backreferences for data
extents in 0d928b4 "Show data extent backrefs".

This commit adds handling of the metadata tree block information and
backreferences.

There are two different ways this information can be stored:
* The original way, using a tree_block_info structure inside the extent
  item while setting flags to EXTENT_FLAG_TREE_BLOCK.
* The 'skinny' way, which is a more compact way of encoding all
  necessary information, which does not include the tree_block_info
  struct, and reuses the offset field in the object key for a different
  purpose. See kernel commit 3173a18 for more information about this.

The skinny metadata has its own key type: METADATA_ITEM_KEY.
In order to try keeping the code a bit understandable, I made the
MetaDataItem a separate class, instead of putting more if statements
inside the Extent class itself.

So, if the extent has flags DATA, it can have extent_data_refs and
shared_data_refs, which are inline, or separate objects following the
extent item itself, if there are too many of them to fit inline.

For fat metadata, the key type is also EXTENT_ITEM, but it has the flag
TREE_BLOCK set, because the extent contains btree data. In that case, it
contains the tree_block_info structure and an inline ref:

    extent vaddr 88475074560 length 16384 refs 1 gen 1550691 flags TREE_BLOCK|FULL_BACKREF
        tree block key (8212 DIR_INDEX 24911) level 0 backref root 5
    extent vaddr 88475222016 length 16384 refs 1 gen 1567265 flags TREE_BLOCK|FULL_BACKREF
        tree block key (7567546 INODE_ITEM 0) level 0 backref root 5

I put the tree_block_info inside a TreeBlockInfo object, together with
the additional inline backref.

btrfs-debug-tree shows this as:

    item 128 key (88475074560 EXTENT_ITEM 16384) itemoff 9731 itemsize 51
        extent refs 1 gen 1550691 flags TREE_BLOCK
        tree block key (8212 DIR_INDEX 24911) level 0
        tree block backref root 5
    item 129 key (88475222016 EXTENT_ITEM 16384) itemoff 9680 itemsize 51
        extent refs 1 gen 1567265 flags TREE_BLOCK
        tree block key (7567546 INODE_ITEM 0) level 0
        tree block backref root 5

For skinny metadata, the key type is METADATA_ITEM, and the tree level
and root backref is stored in the MetaDataItem object itself.

    metadata vaddr 14348730368 refs 1 gen 676 flags TREE_BLOCK|FULL_BACKREF tree block skinny level 0 backref root 9
    metadata vaddr 14349385728 refs 1 gen 677 flags TREE_BLOCK|FULL_BACKREF tree block skinny level 0 backref root 257
    metadata vaddr 14350958592 refs 1 gen 692 flags TREE_BLOCK|FULL_BACKREF tree block skinny level 0 backref root 5

btrfs-debug-tree shows this as:

    item 7 key (14348730368 METADATA_ITEM 0) itemoff 15589 itemsize 33
        extent refs 1 gen 676 flags TREE_BLOCK
        tree block skinny level 0
        tree block backref root 9
    item 8 key (14349385728 METADATA_ITEM 0) itemoff 15556 itemsize 33
        extent refs 1 gen 677 flags TREE_BLOCK
        tree block skinny level 0
        tree block backref root 257
    item 9 key (14350958592 METADATA_ITEM 0) itemoff 15523 itemsize 33
        extent refs 1 gen 692 flags TREE_BLOCK
        tree block skinny level 0
        tree block backref root 5

In examples/show_block_group_contents.py, it's only necessary to print
the separate TreeBlockInfo in case of fat metadata.